+++ NEWS.md
@@ -1,3 +1,7 @@
+# tidytracks dev
+
+* Implement `hr_ud_sum()` to sum multiple UDs.
+
 # tidytracks 0.0.1
 
 * Initial public release.